Small modifications it should work fine though

+++ Ball.cpp
@@ -41,10 +41,10 @@ void Ball::reset() {
 	setX(GAME_WIDTH/2 - getWidth() * getScale() / 2);
 	setY(GAME_HEIGHT/2 - getHeight() * getScale() / 2);
 
-	srand(time(NULL));
+	srand((UINT)time(NULL));
 
-	velocity.x = (rand() % 80 + 5)*(rand() % 2 ? 1 : -1);
-	velocity.y = (rand() % 90 + 10)*(rand() % 2 ? 1 : -1);
+	velocity.x = (FLOAT) (rand() % 80 + 5)*(rand() % 2 ? 1 : -1);
+	velocity.y = (FLOAT)(rand() % 90 + 10)*(rand() % 2 ? 1 : -1);
 	Graphics::Vector2Normalize(&velocity);
 	velocity *= speed;
 }
